+++ src/App.js
@@ -1,32 +1,14 @@
 import React from 'react';
+import TodoList from "./TodoList";
+import AddTodoForm from "./AddTodoForm";
 
-let todoList = [
-  {
-    id: 1,
-    title: 'Read assignment',
-  },
-  {
-    id: 2,
-    title: 'Complete assignment',
-  },
-  {
-    id: 3,
-    title: 'Submit assignment',
-  },
-];
 
 function App() {
   return (
     <div>
       <h1 >To Do List</h1>
-      <ul>
-        {todoList.map(item => {
-          return (
-            <li key={item.id}><hr></hr>{item.title}</li>
-            
-          );
-        })}        
-      </ul>
+      <AddTodoForm />
+      <TodoList />      
     </div>
   );
 }
